+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 1.0.0 Launch
 1.0.1 Improving, cleaning and documenting code
-1.0.2 Inserted Class Header
+1.0.2 Add Class Header
+1.0.3 Add Class SceneManager